requirements.txt
flask
flask-cors
python-dotenv
gradio_client
gunicorn